requirements.txt
# Image processing
Pillow==10.4.0
opencv-python==4.10.0.84

# OCR
pytesseract==0.3.10

# AI / ML
torch>=2.2.0
torchvision>=0.17.0
transformers==4.41.2
huggingface-hub==0.23.3
accelerate==0.31.0

# Detection / vision helpers
timm==1.0.7
ultralytics>=8.3.0

# Text similarity
rapidfuzz==3.9.3

# Numerical
numpy==1.26.4
scipy==1.13.1

# Utilities
tqdm==4.66.4
requests==2.32.3
python-dotenv==1.0.1

# API
fastapi
uvicorn[standard]
python-multipart